FIX: no overflow with more than 255 tools ...
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@3607 8aca7d54-2c30-db11-9de9-000461428c89

+++ gui/werkzeug_waehler.h
@@ -34,8 +34,8 @@ private:
 	 * window width in toolboxes
 	 * @author Hj. Malthaner
 	 */
-	uint8 tool_icon_width;
-	uint8 tool_icon_height;
+	uint16 tool_icon_width;
+	uint16 tool_icon_height;
 
 	uint16 tool_icon_disp_start;
 	uint16 tool_icon_disp_end;
